# backend/swagger/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "YBR RestAPI DOCUMENT",
    "version": "1.0.0",
    "description": ""
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
...
    "/api/user/couple": {
      "put": {
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "waitingId": {
                    "example": "any"
                  },
                  "toUID": {
                    "example": "any"
                  },
                  "fromUID": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/user/login": {
      "post": {
        "tags": [
          "user"
        ],
        "description": "YBR에서 제공하는 일반 로그인 API",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "params",
            "in": "params",
            "schema": {
              "type": "object",
              "properties": {
                "params": {
                  "type": "object",
                  "properties": {
                    "userId": {
                      "type": "string",
                      "example": "primero"
                    },
                    "password": {
                      "type": "string",
                      "example": "1234"
                    },
                    "rememberMe": {
                      "type": "boolean",
                      "example": true
                    }
                  }
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "params": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "in": "header",
        "bearerFormat": "JWT"
      }
    }
  }
}